+++ packages/headless/src/api/analytics/search-analytics.ts
@@ -43,6 +43,7 @@ export class SearchAnalyticsProvider
   extends BaseAnalyticsProvider
   implements SearchPageClientProvider
 {
+  private static fallbackPipelineName = 'default';
   constructor(private state: StateNeededBySearchAnalyticsProvider) {
     super(state);
   }
@@ -53,7 +54,9 @@ export class SearchAnalyticsProvider
 
   public getPipeline() {
     return (
-      this.state.pipeline || this.state.search?.response.pipeline || 'default'
+      this.state.pipeline ||
+      this.state.search?.response.pipeline ||
+      SearchAnalyticsProvider.fallbackPipelineName
     );
   }
 
@@ -75,6 +78,20 @@ export class SearchAnalyticsProvider
       this.state.search?.response.searchUid ||
       getSearchInitialState().response.searchUid
     );
+  }
+
+  public getSplitTestRunName(): string | undefined {
+    return this.state.search?.response.splitTestRun;
+  }
+
+  public getSplitTestRunVersion(): string | undefined {
+    const hasSplitTestRun = !!this.getSplitTestRunName();
+    const effectivePipelineWithSplitTestRun =
+      this.state.search?.response.pipeline ||
+      this.state.pipeline ||
+      SearchAnalyticsProvider.fallbackPipelineName;
+
+    return hasSplitTestRun ? effectivePipelineWithSplitTestRun : undefined;
   }
 
   private mapResultsToAnalyticsDocument() {
@@ -152,6 +169,12 @@ export const configureAnalytics = ({
 
 export const getVisitorID = () =>
   new CoveoAnalyticsClient({}).getCurrentVisitorId();
+
+export const clearAnalyticsClient = () => {
+  const client = new CoveoAnalyticsClient({});
+  client.clear();
+  client.deleteHttpOnlyVisitorId();
+};
 
 export const historyStore = new history.HistoryStore();