+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.13...@coveo/atomic-react@1.15.14) (2022-07-15)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.12...@coveo/atomic-react@1.15.13) (2022-07-15)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.15.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.11...@coveo/atomic-react@1.15.12) (2022-07-14)
 
 **Note:** Version bump only for package @coveo/atomic-react